requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
httpx==0.27.0
pydantic==2.7.1
gunicorn==22.0.0
aiofiles==23.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
